+++ instrumentation/opentelemetry-instrumentation-botocore/tests/test_botocore_instrumentation.py
@@ -27,7 +27,12 @@ from moto import (  # pylint: disable=import-error
 )
 
 from opentelemetry import trace as trace_api
-from opentelemetry.context import attach, detach, set_value
+from opentelemetry.context import (
+    _SUPPRESS_HTTP_INSTRUMENTATION_KEY,
+    attach,
+    detach,
+    set_value,
+)
 from opentelemetry.instrumentation.botocore import BotocoreInstrumentor
 from opentelemetry.instrumentation.utils import _SUPPRESS_INSTRUMENTATION_KEY
 from opentelemetry.propagate import get_global_textmap, set_global_textmap
@@ -325,6 +330,17 @@ class TestBotocoreInstrumentor(TestBase):
         finally:
             detach(token)
         self.assertEqual(0, len(self.get_finished_spans()))
+
+    @mock_xray
+    def test_suppress_http_instrumentation_xray_client(self):
+        xray_client = self._make_client("xray")
+        token = attach(set_value(_SUPPRESS_HTTP_INSTRUMENTATION_KEY, True))
+        try:
+            xray_client.put_trace_segments(TraceSegmentDocuments=["str1"])
+            xray_client.put_trace_segments(TraceSegmentDocuments=["str2"])
+        finally:
+            detach(token)
+        self.assertEqual(2, len(self.get_finished_spans()))
 
     @mock_s3
     def test_request_hook(self):