+++ instrumentation/opentelemetry-instrumentation-urllib3/tests/test_urllib3_integration.py
@@ -20,10 +20,10 @@ import urllib3
 import urllib3.exceptions
 
 from opentelemetry import context, trace
-from opentelemetry.instrumentation.urllib3 import (
-    _SUPPRESS_HTTP_INSTRUMENTATION_KEY,
-    URLLib3Instrumentor,
-)
+
+# FIXME: fix the importing of this private attribute when the location of the _SUPPRESS_HTTP_INSTRUMENTATION_KEY is defined.
+from opentelemetry.context import _SUPPRESS_HTTP_INSTRUMENTATION_KEY
+from opentelemetry.instrumentation.urllib3 import URLLib3Instrumentor
 from opentelemetry.instrumentation.utils import _SUPPRESS_INSTRUMENTATION_KEY
 from opentelemetry.propagate import get_global_textmap, set_global_textmap
 from opentelemetry.semconv.trace import SpanAttributes